This make the fileupload in line with the rest of the project's style of code
