+++ src/api/upload.js
@@ -1,49 +1,35 @@
 import resource from 'resource-router-middleware'
+import formidable from 'formidable'
+import * as fs from 'fs'
+import parse_csv from '../services/upload/parseCSV'
+import path from 'path'
 
 
-'use strict';
+export default ({config, db}) => resource({
 
-var formidable = require('formidable');
-var path = require('path');
-var fs = require('fs');
-//var csv_parser = require('./csvParser');
-var parse_csv = require('../services/csvUpload/testing_csvjson');
+    /** POST - Create a new upload */
+    create(req, res) {
 
+        let form = new formidable.IncomingForm();
+        
+        form.multiples = true;
+        form.uploadDir = path.join(__dirname, '../services/upload/uploads');
 
-export default ({ config, db }) => resource({
-	
-	/** POST - Create a new upload */
-	create(req, res) {
-
-		var form = new formidable.IncomingForm();
-    
-    	form.multiples = true;
-    	form.uploadDir = path.join(__dirname, '../services/csvUpload/uploads');
-        console.log(form.uploadDir);
-      	form.on('file', function rename_file(field, file) {
-        	fs.renameSync(file.path, path.join(form.uploadDir, file.name));
-            parse_csv(null,path.join(form.uploadDir, file.name));
-    	});
-
-    	form.on('error', function(err) {
-        	console.log('An error has occured: \n' + err);
-    	});
-
-    	form.on('end', function() {
-        	res.end('success');
-    	});
-    	form.parse(req);
-		}
-})
-
-/*
-function pass_csv(form, cb) {
-    reutrn new Promise (function (fulfill, reject){
-            fs.rename(file.path, path.join(form.uploadDir, file.name), parse_csv(null, file.path){;
-            var data = cb('../services/csvUpload/uploads'+file.name);
-            //var csv_parser = require('./csvParser');
-            //var data = csv_parser('./uploads/'+file.name);
-            console.log(data);
+        form.on('file', function rename_file(field, file) {
+        fs.renameSync(file.path, path.join(form.uploadDir, file.name))
+        
+        parse_csv(path.join(form.uploadDir, file.name),db);
         });
+
+        form.on('error', function(err) {
+            console.error(err);
+            res.sendStatus(404);
+        });
+
+        form.on('end', function() {
+            res.end('success');
+        });
+
+        form.parse(req);
     }
-}*/
+})
