+++ src/config.json
@@ -1,8 +1,5 @@
 {
 	"port": 8080,
 	"bodyLimit": "100kb",
-	"corsHeaders": ["Link"],
-    "mysql_user": "root",
-    "mysql_password": "tracysell",
-    "mysql_host": "104.197.192.209"
+	"corsHeaders": ["Link"]
 }
