--- src/uploadCSV.html
@@ -1,30 +1,0 @@
-<html>
-<head>
-<title>Mmmm</title>
-    <script   src="https://code.jquery.com/jquery-1.12.4.min.js"   integrity="sha256-ZosEbRLbNQzLpnKIkEdrPv7lOy9C27hHQ+Xp8a4MxAQ="   crossorigin="anonymous"></script>
-    <script   src="https://code.jquery.com/jquery-migrate-1.2.1.min.js"   integrity="sha256-HmfY28yh9v2U4HfIXC+0D6HCdWyZI42qjaiCFEJgpo0="   crossorigin="anonymous"></script>
-    <!-- Latest compiled and minified CSS -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css"
-          integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
-
-    <!-- Optional theme -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap-theme.min.css"
-          integrity="sha384-fLW2N01lMqjakBkx3l/M9EahuwpSfeNvV63J5ezn3uZzapT0u7EYsXMjQV+0En5r" crossorigin="anonymous">
-    <!-- Latest compiled and minified JavaScript -->
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"
-            integrity="sha384-0mSbJDEHialfmuBBQP6A4Qrprq5OVfW37PRR3j5ELqxss1yVqOtnepnHVP9aJ7xS"
-            crossorigin="anonymous"></script>
-
-</head>
-<body>
-<div class="container-fluid center-block col-md-4 col-md-offset-4" style="margin-top: 100px; float: none;">
-    <form id="entryForm" name="entryForm" action="http://localhost:8080/api/upload" enctype="multipart/form-data" method="post">
-        <div class="form-group">
-            <label>File to Upload</label>
-            <input class="form-control" type="file" name="students" id="students" required />
-            <button type="submit" name="submit" id="submit" class="btn btn-default">Submit</button>
-        </div>
-    </form>
-</div>
-</body>
-</html>
